requirements.txt
pytest==7.4.3
selenium==4.15.2
webdriver-manager==4.0.1
allure-pytest==2.13.2
